bar-log/requirements.txt
streamlit>=1.31
mysql-connector-python>=8.2
pandas>=2.1
